Move exception to its own file
